feat: Use monolithic helper for split tests (temp)
