--- test-tape/mocha-source-split/helpers/index.js
@@ -1,19 +1,0 @@
-// test-tape/mocha-source-split/helpers/index.js
-
-// Re-export everything from the split modules
-
-const config = require('./config');
-const request = require('./request');
-const tables = require('./tables');
-const batch = require('./batch');
-const assertions = require('./assertions');
-const setup = require('./setup');
-
-module.exports = {
-  ...config,
-  ...request,
-  ...tables,
-  ...batch,
-  ...assertions,
-  ...setup,
-};
